Extract some JVM-specific member functions to extension functions
